Switch to webpack
Rollup just isn't mature enough. In particular:

* It generates some cyclic dependency errors when importing chai
* The karma-rollup-preprocessor doesn't seem to map source file names
  propery (I worked around this locally by adding the following line to
  `createPreprocessor`):

```diff
          const { code, map } = result;
          const { sourcemap } = options.output;

          file.sourceMap = map;
+         file.path = map.file;

          const processed =
            sourcemap === "inline"
              ? code + `\n//# sourceMappingURL=${map.toUrl()}\n`
              : code;
```

* It has create trouble with fetch-mock. In particular it seems to need
  rollup-plugin-node-builtins (for query string) and then when it gets
  into the URL handling part of the code (using the WHATWG URL package)
  it somehow ends up in the node.js code path despite running in the
  browser which just means that everything breaks.

In the end it was just way too much work to fix all these bugs.

This patch switches us to Webpack which brings a new set of problems.
Namely that ts-loader doesn't know how to deal with svelte components.

Basically, at the time of writing (2019-07-13) Svelte's support for TS
is woeful. I'd love to use it but it's just not ready yet.

  https://github.com/sveltejs/svelte/issues/1639