Remove no-longer-needed superstatic setup
We originally introduced this since it seemed like running a separate
web server would be better (I was having trouble with fetch-mock) but
for the time being that no longer seems necessary.

I checked it in, however, simply because we might well want to go back
to this approach at some pointer.